% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfvector.R
\name{as.unfvector}
\alias{as.unfvector}
\title{UNF Vector Representation}
\usage{
as.unfvector(x, ...)
}
\arguments{
\item{x}{A vector to be coerced to a character string representation according to the UNF specification.}

\item{\ldots}{Additional arguments passed to methods.}
}
\value{
A character string with class \dQuote{unfvector} manipulated to follow the UNF specification. These are used internally by \code{\link{unf6}}.
}
\description{
Standardize a vector according to UNF specifications
}
\details{
The UNF specifications describes how to coerce all R data types to a standardized character representation. This S3 method exposes that coercion functionality.
}
\author{
Thomas J. Leeper (\email{thosjleeper@gmail.com})
}
\seealso{
\code{\link{unf}}, \code{\link{unf6}}, \code{\link{\%unf\%}}
}

