\name{unf}
\alias{unf}

\title{Multidimensional Uniformity Test}
\description{
Return the p-value of MST-based test of Smith and Jain
}
\usage{
unf(x)
}

\arguments{
  \item{x}{A \code{matrix} of real numbers, the columns of which
  have to be tested for uniformity.}
}

\details{ 
  See Smith and Jain (1984) for detail.
}

\value{
  The p-value of the test.
}

\author{Ben Pfaff}
\keyword{multivariate}
